# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VenusConnect - WhatsApp API Gateway",
    "description": "REST API untuk mengirim dan menerima pesan WhatsApp menggunakan Baileys library.\n\n## Authentication\n\nSemua endpoint (kecuali `/health` dan `/api/auth/*`) memerlukan API key di header:\n\n```\nx-api-key: your-api-key-here\n```\n\n## Rate Limiting\n\n- **100 requests/minute** per IP\n- Jika limit terlampaui, response akan error 429\n\n## Media Support\n\nMedia dapat dikirim dalam 3 format:\n- **URL**: `https://example.com/image.jpg`\n- **Local Path**: `/path/to/file.pdf`\n- **Base64**: `data:image/jpeg;base64,/9j/4AAQ...`\n\n## Webhook Events\n\nJika webhook URL dikonfigurasi, events berikut akan dikirim:\n- `message.received` - Pesan masuk\n- `message.status` - Status pesan (sent/delivered/read)\n- `presence.update` - Online/offline/typing",
    "version": "1.0.0",
    "contact": {
      "name": "VenusConnect Support",
      "email": "support@venusconnect.io",
      "url": "https://venusconnect.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    },
    {
      "url": "https://api.venusconnect.io",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register new user",
        "description": "Create a new user account (public endpoint)",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "User login",
        "description": "Authenticate and get API key",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/session": {
      "get": {
        "tags": ["Session"],
        "summary": "List all sessions",
        "description": "Get all WhatsApp sessions for current user",
        "operationId": "listSessions",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "Sessions list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Session" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Session"],
        "summary": "Create new session",
        "description": "Create a new WhatsApp session and get QR code",
        "operationId": "createSession",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/SessionStatus" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{sessionId}": {
      "get": {
        "tags": ["Session"],
        "summary": "Get session status",
        "operationId": "getSessionStatus",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" }
        ],
        "responses": {
          "200": {
            "description": "Session status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/SessionStatus" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Session"],
        "summary": "Delete session",
        "description": "Logout and delete WhatsApp session",
        "operationId": "deleteSession",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" }
        ],
        "responses": {
          "200": {
            "description": "Session deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/session/{sessionId}/send": {
      "post": {
        "tags": ["Messaging"],
        "summary": "Send message",
        "description": "Send text and/or media to individual contact",
        "operationId": "sendMessage",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/SessionId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SendMessageRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message sent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication"
      }
    },
    "parameters": {
      "SessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "WhatsApp session ID"
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "minLength": 3, "example": "johndoe" },
          "email": { "type": "string", "format": "email", "example": "john@example.com" },
          "password": { "type": "string", "minLength": 6, "example": "secret123" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "username": { "type": "string" },
              "role": { "type": "string", "enum": ["admin", "user"] },
              "api_key": { "type": "string" }
            }
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "session_id": { "type": "string" },
          "status": { "type": "string", "enum": ["connecting", "connected", "disconnected"] },
          "phone_number": { "type": "string" },
          "name": { "type": "string" },
          "connected": { "type": "boolean" },
          "hasQr": { "type": "boolean" }
        }
      },
      "SessionStatus": {
        "type": "object",
        "properties": {
          "session_id": { "type": "string" },
          "status": { "type": "string" },
          "connected": { "type": "boolean" },
          "qr": { "type": "string", "nullable": true, "description": "Base64 QR code" },
          "message": { "type": "string" }
        }
      },
      "CreateSessionRequest": {
        "type": "object",
        "properties": {
          "session_id": { "type": "string", "description": "Custom session ID (optional)" },
          "webhook_url": { "type": "string", "format": "uri", "description": "Webhook URL for events" }
        }
      },
      "SendMessageRequest": {
        "type": "object",
        "required": ["to"],
        "properties": {
          "to": { "type": "string", "description": "Phone number (e.g., 6281234567890)", "example": "6281234567890" },
          "message": { "type": "string", "description": "Text message" },
          "media": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MediaItem" }
          }
        }
      },
...
      "MediaItem": {
        "type": "object",
        "required": ["type", "data"],
        "properties": {
          "type": { "type": "string", "enum": ["image", "video", "document", "audio", "sticker"] },
          "data": { "type": "string", "description": "URL, local path, or base64" },
          "caption": { "type": "string" },
          "filename": { "type": "string" },
          "mimetype": { "type": "string" }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "messageId": { "type": "string" }
            }
          }
        }
      },
...
    }
  }
}